Moved TagsEditor to tags directory
