+++ src/routes/features/groups/[id]/edit/+page.svelte
@@ -8,7 +8,7 @@
     import TextField from "$components/ui/forms/TextField.svelte";
     import Menu from "$components/ui/menu/Menu.svelte";
     import MenuItem from "$components/ui/menu/MenuItem.svelte";
-    import TagsEditor from "$components/web-components/TagsEditor.svelte";
+    import TagsEditor from "$components/tags/TagsEditor.svelte";
     import TagGroup from "$entities/TagGroup.ts";
     import {tagGroupsStore} from "$stores/tag-groups-store.js";
